+++ src/config/titulo.js
@@ -1,2 +1,1 @@
-module.exports =
-  'Conceptualización 3D'
+module.exports = 'Conceptualización 3D'